% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorbitops-package.R
\docType{package}
\name{vectorbitops-package}
\alias{vectorbitops}
\alias{vectorbitops-package}
\title{vectorbitops: Vector Bitwise Operations}
\description{
A tool for fast, efficient bitwise operations along the elements within a vector. Provides such functionality for AND, OR and XOR, as well as infix operators for all of the binary bitwise operations.
}
\author{
\strong{Maintainer}: Samuel Sapire \email{sapires@protonmail.com} [copyright holder]

}
\keyword{internal}
